+++ resume.sh
@@ -2,6 +2,7 @@
 #
 # Resumes an already running remote sbatch job.
 # Sample usage: bash resume.sh
+
 if [ ! -f params.sh ]
 then
     echo "Need to configure params before first run, run setup.sh!"
@@ -9,5 +10,7 @@ then
 fi
 source params.sh
 
-MACHINE=`ssh sherlock squeue --name=$NAME --user=$USERNAME -o "%N" -h`
-ssh -L $PORT:localhost:$PORT sherlock ssh -L $PORT:localhost:$PORT -N $MACHINE &
+# The user is required to specify port
+
+MACHINE=`ssh ${RESOURCE} squeue --name=$NAME --user=$USERNAME -o "%N" -h`
+ssh -L $PORT:localhost:$PORT ${RESOURCE} ssh -L $PORT:localhost:$PORT -N $MACHINE &